+++ src/ui/wmap_ui.h
@@ -20,7 +20,7 @@ void sub_563770(int direction);
 void sub_564000(int a1);
 void sub_564AF0(int64_t a1);
 bool wmap_ui_bkg_process_callback(TimeEvent* timeevent);
-void sub_564F60(long long a1, long long a2);
+void sub_564F60(int64_t a1, int64_t a2);
 void sub_566CC0(int64_t* location_ptr);
 
 #endif /* ARCANUM_UI_WMAP_UI_H_ */
